Mixer: Put time limit to mixer rectangle
The mixer rectangle halo seems to cause a noticeable performance
overhead when shown on screen, so limiting the time it appears on
screen should be better than making it appear indefinitely by default.
This commit also adds an option in the user config file to adjust
the time the rectangle appears on screen.